chore: move fakeVatAdmin to tools/ but leave export in contractFacet/
not a breaking change. Also change the references.
