+++ packages/deploy-script-support/test/unitTests/test-install.js
@@ -3,7 +3,7 @@
 import { test } from '@agoric/zoe/tools/prepare-test-env-ava';
 
 import { makeZoe } from '@agoric/zoe';
-import fakeVatAdmin from '@agoric/zoe/src/contractFacet/fakeVatAdmin';
+import fakeVatAdmin from '@agoric/zoe/tools/fakeVatAdmin';
 import { makeBoard } from '@agoric/vats/src/lib-board';
 import bundleSource from '@agoric/bundle-source';
