+++ packages/zoe/test/unitTests/contractSupport/test-depositTo.js
@@ -8,7 +8,7 @@ import bundleSource from '@agoric/bundle-source';
 
 import { setup } from '../setupBasicMints';
 import { makeZoe } from '../../..';
-import { makeFakeVatAdmin } from '../../../src/contractFacet/fakeVatAdmin';
+import { makeFakeVatAdmin } from '../../../tools/fakeVatAdmin';
 import { depositToSeat } from '../../../src/contractSupport/zoeHelpers';
 import { makeOffer } from '../makeOffer';
