+++ packages/zoe/test/unitTests/contractSupport/test-offerTo.js
@@ -8,7 +8,7 @@ import bundleSource from '@agoric/bundle-source';
 
 import { setup } from '../setupBasicMints';
 import { makeZoe } from '../../..';
-import { makeFakeVatAdmin } from '../../../src/contractFacet/fakeVatAdmin';
+import { makeFakeVatAdmin } from '../../../tools/fakeVatAdmin';
 import {
   offerTo,
   assertProposalShape,
