+++ packages/zoe/test/unitTests/contracts/test-escrowToVote.js
@@ -9,7 +9,7 @@ import { E } from '@agoric/eventual-send';
 
 import { makeZoe } from '../../../src/zoeService/zoe';
 import { setup } from '../setupBasicMints';
-import fakeVatAdmin from '../../../src/contractFacet/fakeVatAdmin';
+import fakeVatAdmin from '../../../tools/fakeVatAdmin';
 
 const contractRoot = `${__dirname}/escrowToVote`;
