+++ packages/zoe/test/unitTests/contracts/test-mintPayments.js
@@ -6,7 +6,7 @@ import { test } from '@agoric/zoe/tools/prepare-test-env-ava';
 import bundleSource from '@agoric/bundle-source';
 import { E } from '@agoric/eventual-send';
 import { makeIssuerKit, amountMath } from '@agoric/ertp';
-import fakeVatAdmin from '../../../src/contractFacet/fakeVatAdmin';
+import fakeVatAdmin from '../../../tools/fakeVatAdmin';
 
 // noinspection ES6PreferShortImport
 import { makeZoe } from '../../../src/zoeService/zoe';
