+++ packages/zoe/test/unitTests/contracts/test-secondPriceAuction.js
@@ -12,7 +12,7 @@ import { makeZoe } from '../../../src/zoeService/zoe';
 import buildManualTimer from '../../../tools/manualTimer';
 import { setup } from '../setupBasicMints';
 import { setupMixed } from '../setupMixedMints';
-import fakeVatAdmin from '../../../src/contractFacet/fakeVatAdmin';
+import fakeVatAdmin from '../../../tools/fakeVatAdmin';
 
 const secondPriceAuctionRoot = `${__dirname}/../../../src/contracts/auction/secondPriceAuction`;
