+++ packages/zoe/test/unitTests/contracts/test-throwInOfferHandler.js
@@ -7,7 +7,7 @@ import bundleSource from '@agoric/bundle-source';
 
 import { E } from '@agoric/eventual-send';
 import { makeZoe } from '../../../src/zoeService/zoe';
-import fakeVatAdmin from '../../../src/contractFacet/fakeVatAdmin';
+import fakeVatAdmin from '../../../tools/fakeVatAdmin';
 
 const contractRoot = `${__dirname}/throwInOfferHandler`;
