+++ packages/zoe/test/unitTests/test-makeKind.js
@@ -7,7 +7,7 @@ import bundleSource from '@agoric/bundle-source';
 
 import { E } from '@agoric/eventual-send';
 import { makeZoe } from '../../src/zoeService/zoe';
-import fakeVatAdmin from '../../src/contractFacet/fakeVatAdmin';
+import fakeVatAdmin from '../../tools/fakeVatAdmin';
 
 const root = `${__dirname}/../minimalMakeKindContract`;
